requirements.txt
pip~=23.2.1
Django~=5.1.2
asgiref~=3.8.1
sqlparse~=0.5.1
colorama~=0.4.6
click~=8.1.7
djangorestframework~=3.15.2